-add two find solution, recursive and non recursive. -optimized the representation of a hole in memory, instead of use one int or byte to represent a hole, I used ONE bit for each hole so on a byte I able to represent eight ribbon hole.

+++ RibbonsColor/generator/IGenerateRibbons.cs
@@ -1,14 +1,9 @@
 ﻿using RibbonsColor.model;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace RibbonsColor.generator
 {
     public interface IGenerateRibbons
     {
-        public List<IRibbonModel> Generate(int maxHoles);
+        public List<IRibbonModel> Generate(long maxHoles);
     }
 }
